fix: better way to track invalid table cell paste (#1106)
